chore: Update database connection in middleware tests and remove unused imports in handlers tests
